fix: Return empty title if readable fails
Some web pages (e.g. tweets on twitter.com) don't have a content, and
readability-cli fails on such pages. In this case, it would be better to return
nil as the title rather than aborting the entire process, because it is still
possible to produce an Org link without a title.