fix: Avoid running readability-cli for parsing canonical URL
readability-cli often fails with the following error message:

> Not sure if this document should be processed. Not processing

This is often because the target page is supposed to be consumed only via a web
browser and has no static content. It is still possible to scrape metadata from
those pages. It is often faster to prevent running of the command line.